Move autolykos_pow_scheme to ergo-chain-types
