+++ ergo-nipopow/src/lib.rs
@@ -18,11 +18,10 @@
 #![deny(clippy::unimplemented)]
 #![deny(clippy::panic)]
 
-mod autolykos_pow_scheme;
 mod nipopow_algos;
 mod nipopow_proof;
 mod nipopow_verifier;
 
-pub use nipopow_algos::{decode_compact_bits, NipopowAlgos, INTERLINK_VECTOR_PREFIX};
+pub use nipopow_algos::{NipopowAlgos, INTERLINK_VECTOR_PREFIX};
 pub use nipopow_proof::{NipopowProof, NipopowProofError, PoPowHeader};
 pub use nipopow_verifier::NipopowVerifier;
